+++ src/app.py
@@ -1,26 +1,20 @@
 import os
+from typing import Dict
 import requests
 from dotenv import load_dotenv
 from api_client import ApiClient
 from models import *
-from inventory import InventoryManager
-from processing_queue import KitProcessingQueue
 
 from dataclasses import dataclass
 from state import State
 from context import Context
 from decision_maker import DecisionMaker
+from utils import encode_time
 
 DATA_DIR = os.path.join(os.path.dirname(os.path.dirname(os.path.abspath(__file__))), "data")
 
 @dataclass
 class App:
-    aircraft_dict: Dict[str, Aircraft] = field(default_factory=dict)
-    airport_dict: Dict[str, Airport] = field(default_factory=dict)
-    flights_dict: Dict[str, Flight] = field(default_factory=dict)
-    inventory_manager: Optional[InventoryManager] = None
-    processing_queue: Optional[KitProcessingQueue] = None
-
     context: Context = field(default_factory=Context)
     state: State = field(init=False)
     decisionMaker: DecisionMaker = field(init=False)
@@ -41,136 +35,29 @@ class App:
         self.client = ApiClient(base_url=BASE_URL, api_key=API_KEY)
 
     def run(self):
+        # connect the api
         self.connect_api()
-
-        # Example of how you might use the parse methods
-        parser = Parser()
-
-        self.aircraft_dict = parser.parse_aircraft('../data/aircraft_types.csv')
-        self.airport_dict = parser.parse_airports('../data/airports_with_stocks.csv')
-
-        self.inventory_manager = InventoryManager.from_airports(self.airport_dict)
-        self.processing_queue = KitProcessingQueue()
-
-        if self.inventory_manager:
-            print("Initial inventories: ")
-            print(self.inventory_manager.snapshot())
-
-        #print(self.aircraft_dict)
-        #print(self.airport_dict)
-
-    def _processing_times_for_airport(self, airport_code: str) -> Dict[str, int]:
-        airport = self.airport_dict.get(airport_code)
-        if not airport:
-            return {}
-        return {
-            "first": airport.first_processing_time,
-            "business": airport.business_processing_time,
-            "premiumEconomy": airport.premium_economy_processing_time,
-            "economy": airport.economy_processing_time,
-        }
-
-    def enqueue_kits_for_processing(self, airport_code: str, amounts: Dict[str, int]) -> None:
-        """
-        Call this when kits arrive at an airport; they will become available after processing time.
-        """
-        if not self.processing_queue:
-            return
-        processing_times = self._processing_times_for_airport(airport_code)
-        self.processing_queue.add_batch(airport_code, amounts, processing_times)
-
-    def complete_processing_tick(self) -> None:
-        """
-        Advance processing by one hour and move finished kits into inventory.
-        """
-        if not self.processing_queue or not self.inventory_manager:
-            return
-        violations = self.processing_queue.apply_tick(self.inventory_manager, hours=1)
-        if violations:
-            print("Processing violations:", violations)
-
-    def update_flights(self, response):
-        flights = response["flightUpdates"]
-        for flight_entry in flights:
-            # --- 1. Calculate Time ---
-            departure_days_elapsed = flight_entry["departure"]["day"] - 1
-            arrival_days_elapsed = flight_entry["arrival"]["day"] - 1
-
-            departure_time = encode_time(
-                days=departure_days_elapsed,
-                hours=flight_entry["departure"]["hour"]
-            )
-
-            arrival_time = encode_time(
-                days=arrival_days_elapsed,
-                hours=flight_entry["arrival"]["hour"]
-            )
-
-            # --- 2. Convert Status and Get Aircraft ID ---
-            status_enum = FlightStatus[flight_entry["eventType"]]
-            flight_id = flight_entry["flightId"]
-
-            aircraft_id = ''
-            for id in self.aircraft_dict:
-                if flight_entry["aircraftType"] ==  self.aircraft_dict[id]:
-                    aircraft_id = id
-                    break
-
-            # --- 3. Check for Existing Flight and Update/Create ---
-            if flight_id in self.flights_dict:
-                existing_flight = self.flights_dict[flight_id]
-
-                # Typically, only certain attributes are updated, like status, times, and passengers.
-                # We assume flight number, airport IDs, and aircraft ID remain constant in an update.
-                existing_flight.status = status_enum
-                existing_flight.departure = departure_time
-                existing_flight.arrival = arrival_time
-                existing_flight.passengers = flight_entry["passengers"]
-
-            else:
-                # Flight does not exist: Create and store a new Flight object
-                new_flight = Flight(
-                    status=status_enum,
-                    flight_number=flight_entry["flightNumber"],
-                    flight_id=flight_id,
-                    origin_airport_id=flight_entry["originAirport"],
-                    destination_airport_id=flight_entry["destinationAirport"],
-                    departure=departure_time,
-                    arrival=arrival_time,
-                    passengers=flight_entry["passengers"],
-                    aircraft_id = aircraft_id, # Link to the aircraft ID
-                )
-
-                # Store the new flight
-                self.flights_dict[new_flight.flight_id] = new_flight
-
-    def run(self):
         try:
             self.client.start_session()
-            end_time = 24*1 + 5
             
+            end_time = 24*5
             # main loop for every hour
             while self.state.time < end_time:
-                # advance processing queue for one hour; finished kits return to inventory
-                self.complete_processing_tick()
-                
                 # 1. make a descision
                 decision = self.decisionMaker.make_decision(self.state)
                 
                 # 2. send the decision and get the next round
                 response = self.client.play_round(decision)
-                
-                print(response) # debug response
 
                 # 3. update the state with the next round
                 self.state.update_state(response)
                 
-
         except requests.exceptions.HTTPError as e:
             print(f"HTTP Error: {e.response.status_code} {e.response.reason}")
             print(e.response.json())
         except Exception as e:
             print(f"Error parsing airports: {e}")
         finally:
-            if stop_time != 24 * 30:
+            MAX_END_TIME= 24*30
+            if self.state.time < MAX_END_TIME:
                 self.client.end_session()
